+++ src/main.rs
@@ -4,6 +4,7 @@
 #![forbid(unsafe_code)]
 
 mod client;
+mod config;
 mod entry;
 mod message;
 mod opt;
@@ -16,11 +17,6 @@ use opt::Opt;
 use structopt::StructOpt;
 
 fn main() {
-    if std::env::var_os("RUST_LOG").is_none() {
-        std::env::set_var("RUST_LOG", "info");
-    }
-    pretty_env_logger::init();
-
     let opt = Opt::from_args();
 
     if let Err(err) = opt.run() {